+++ doc/changelog.rst
@@ -6,6 +6,8 @@ Version 0.7.4
 
 * metadata: add ``mem_max_rss`` and ``python_hash_seed``
 * add :func:`perf.python_implementation` and :func:`perf.python_has_jit`
+* TextRunner now recalibrates the benchmark if Python has a JIT compiler.
+  Add ``warmup_loops`` metadata in this case.
 
 Version 0.7.3 (2016-08-17)
 --------------------------